+++ server/context.js
@@ -1,0 +1,7 @@
+export const context = {
+  LOCKED: false,
+  ONLINE: 0,
+  PASSWORD: "changeme", // Should probably move this to a .env file later and make it unchangeable
+  MAX_ROOMS: 25,
+  BLOCKED: ["example"],
+}
