+++ src/main.js
@@ -1,7 +1,7 @@
 import { createApp } from "vue"
 import App from "./App.vue"
 
-import "./index.css"
+import "./assets/index.css"
 
 const app = createApp(App)
